Check an app's origin first for its runtime
Currently when searching for a remote to provide the runtime for an app,
we search remotes in priority order. This commit makes it so we search
the remote providing the app first, then search the rest in priority
order. This behavior just seems logical, but the impetus for this change
was also to keep a unit test passing in eos-updater[1] after a port to
FlatpakTransaction.

This commit maintains the behavior of ignoring nodeps, noenumerate, and
disabled remotes when searching for runtime dependencies.

In practice it's usually true that only one remote provides a runtime
and priorities aren't set at all, so this is an edge case that probably
doesn't come up much.

A unit test is included.

[1] https://github.com/endlessm/eos-updater/blob/eede0a8b9cb7c9cbb8539cfcaad9ce18913dcdd8/tests/test-update-install-flatpaks.c#L1919